TimeStepManager was renamed to WaterBoundaryIntegralCalculator to reflect the very specialized nature of that class.
The calculator now derives from a base class called "AutonomousProblem" to represent all autonomous problems (i.e. independent of time). This has for aim to have solvers that can take an "AutonomousProblem" (perhaps could be renamed to AutonomousProblemCalculator) and apply different maths to solve the ODE problem.

The aim of these changes is to pave the way to separating the problem logic (i.e. the water boundary integral) from the ODE solvers in order to improve modularity of the code and improve testability of the code as well. So different ODE solvers can be tested independent of the problem.