Enhance Runge-Kutta stepper with data logging
Modified `AutonomousRungeKuttaStepperBase` to include a `DataLogger` for improved data collection during steps. Updated `runStep` to accept a step index for conditional logging. Introduced a new `DataLogger` class for asynchronous data copying and added a `PingPongBuffer` class for buffer management. Updated `runTimeStep` to initialize and utilize the `DataLogger`, and enhanced plotting to visualize logged data over time.

The enhanced class can now log the state variables every x number of steps using a separate stream. This allows for data to be transfered without stopping the calculations. Indeed, the RK4 class will wait for a copy to have started and check that the copy is done before modifying the devY0 buffer. This is done to avoid overwritting the data before it's logged. The data copy will only happend when the default stream (computation stream) is done with all the operations by using the event readyToCopy. Indeed, the copyStream will wait for the event readyToCopy before running a memcpyAsync.