Add CuSuperHelium.App project and update simulation logic
Created a new project file `CuSuperHelium.App.vcxproj` with configurations for Debug and Release builds. Added a new source file `kernel.cu` implementing the `modeSum` function and main simulation logic. Updated the solution file to include the new project. Changed the output type of `CuSuperHelium.vcxproj` from `Application` to `DynamicLibrary` and updated the `LanguageStandard` to `stdcpp23`. Modified source file inclusions and added new header files `Export.cuh` and `Export.cu` for dispersion testing. Enhanced `SimulationFunctions.cuh`, `SimulationOptions.h`, and `runSimulation` for improved functionality and error handling. Updated `ValueLogger` to prevent logging of invalid values and added new kernel functions in `createM.cuh`.

The goal of this update is to a) make it easy to change solvers with templates, b) separate the main files, and the console application used for running the simulation so that the main files can be compiled into a DLL that can be called from Python.